Rename section components
