+++ src/components/LPSection.tsx
