+++ src/components/StockSection.tsx
